refactor: move QueryBuilder trait and StreamEventSink abstraction to windmill-ai
