+++ backend/windmill-ai/src/lib.rs
@@ -4,4 +4,5 @@ pub mod ai_cache;
 pub mod ai_google;
 pub mod ai_providers;
 pub mod ai_types;
+pub mod query_builder;
 pub mod types;
