+++ backend/windmill-worker/src/ai/providers/other.rs
@@ -6,7 +6,7 @@ use windmill_common::{client::AuthedClient, error::Error};
 
 use crate::ai::{
     image_handler::prepare_messages_for_api,
-    query_builder::{BuildRequestArgs, ParsedResponse, QueryBuilder, StreamEventProcessor},
+    query_builder::{BuildRequestArgs, ParsedResponse, QueryBuilder, StreamEventSink},
     sse::{OpenAISSEParser, SSEParser},
     types::*,
     utils::should_use_structured_output_tool,
@@ -201,9 +201,9 @@ impl QueryBuilder for OtherQueryBuilder {
     async fn parse_streaming_response(
         &self,
         response: reqwest::Response,
-        stream_event_processor: StreamEventProcessor,
+        stream_event_sink: Box<dyn StreamEventSink>,
     ) -> Result<ParsedResponse, Error> {
-        let mut openai_sse_parser = OpenAISSEParser::new(stream_event_processor);
+        let mut openai_sse_parser = OpenAISSEParser::new(stream_event_sink);
         openai_sse_parser.parse_events(response).await?;
 
         let OpenAISSEParser {
