+++ backend/windmill-worker/src/ai/sse.rs
@@ -15,7 +15,7 @@ use windmill_common::{
 };
 
 use crate::ai::{
-    query_builder::StreamEventProcessor,
+    query_builder::StreamEventSink,
     types::{StreamingEvent, UrlCitation},
 };
 
@@ -109,13 +109,13 @@ pub struct OpenAISSEParser {
     pub accumulated_content: String,
     pub accumulated_tool_calls: HashMap<i64, OpenAIToolCall>,
     pub events_str: String,
-    pub stream_event_processor: StreamEventProcessor,
+    pub stream_event_processor: Box<dyn StreamEventSink>,
     /// Token usage from final chunk (when stream_options.include_usage is true)
     pub usage: Option<OpenAIChatUsage>,
 }
 
 impl OpenAISSEParser {
-    pub fn new(stream_event_processor: StreamEventProcessor) -> Self {
+    pub fn new(stream_event_processor: Box<dyn StreamEventSink>) -> Self {
         Self {
             accumulated_content: String::new(),
             accumulated_tool_calls: HashMap::new(),
@@ -305,7 +305,7 @@ pub struct AnthropicSSEParser {
     pub accumulated_content: String,
     pub accumulated_tool_calls: HashMap<i64, OpenAIToolCall>,
     pub events_str: String,
-    pub stream_event_processor: StreamEventProcessor,
+    pub stream_event_processor: Box<dyn StreamEventSink>,
     /// Track content block types by index
     content_blocks: HashMap<usize, ContentBlockState>,
     /// Collected URL citation annotations from web search
@@ -317,7 +317,7 @@ pub struct AnthropicSSEParser {
 }
 
 impl AnthropicSSEParser {
-    pub fn new(stream_event_processor: StreamEventProcessor) -> Self {
+    pub fn new(stream_event_processor: Box<dyn StreamEventSink>) -> Self {
         Self {
             accumulated_content: String::new(),
             accumulated_tool_calls: HashMap::new(),
@@ -471,7 +471,7 @@ pub struct GeminiSSEParser {
     pub accumulated_content: String,
     pub accumulated_tool_calls: HashMap<i64, OpenAIToolCall>,
     pub events_str: String,
-    pub stream_event_processor: StreamEventProcessor,
+    pub stream_event_processor: Box<dyn StreamEventSink>,
     tool_call_index: i64,
     pub annotations: Vec<UrlCitation>,
     pub used_websearch: bool,
@@ -479,7 +479,7 @@ pub struct GeminiSSEParser {
 }
 
 impl GeminiSSEParser {
-    pub fn new(stream_event_processor: StreamEventProcessor) -> Self {
+    pub fn new(stream_event_processor: Box<dyn StreamEventSink>) -> Self {
         Self {
             accumulated_content: String::new(),
             accumulated_tool_calls: HashMap::new(),
@@ -661,7 +661,7 @@ pub struct OpenAIResponsesSSEParser {
     /// Maps item_id -> accumulated arguments
     tool_call_arguments: HashMap<String, String>,
     pub events_str: String,
-    pub stream_event_processor: StreamEventProcessor,
+    pub stream_event_processor: Box<dyn StreamEventSink>,
     /// Collected URL citation annotations from web search
     pub annotations: Vec<UrlCitation>,
     /// Whether web search was used in this response
@@ -671,7 +671,7 @@ pub struct OpenAIResponsesSSEParser {
 }
 
 impl OpenAIResponsesSSEParser {
-    pub fn new(stream_event_processor: StreamEventProcessor) -> Self {
+    pub fn new(stream_event_processor: Box<dyn StreamEventSink>) -> Self {
         Self {
             accumulated_content: String::new(),
             accumulated_tool_calls: HashMap::new(),
